apply PR comments
